pyproject.toml
[tool.poetry]
name = "bingwebsearch"
version = "0.1.0"
description = "An API wrapper for Bing Web Search using Pydantic."
authors = ["SamEdwardes <edwardes.s@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^1.9.0"
requests = "^2.27.1"
rich = "^11.2.0"
typer = "^0.4.0"

[tool.poetry.dev-dependencies]
python-dotenv = "^0.19.2"
ipykernel = "^6.9.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
